Extract pending migration status hook
